move algorithm.h
