+++ paddle/fluid/operators/optimizers/adamw_op.h
@@ -142,7 +142,7 @@ class SparseAdamWFunctor<T, GPUAdamW, MT> {
 
   inline HOSTDEVICE void operator()(size_t i) const {
     auto row_idx =
-        math::BinarySearch<int64_t>(rows_, row_count_, i / row_numel_);
+        pten::funcs::BinarySearch<int64_t>(rows_, row_count_, i / row_numel_);
     if (lazy_mode_ && row_idx < 0) {
       return;
     } else {
